Move bookmarks DB layer to DAO package
