+++ src/main/java/uk/ac/cam/cl/dtg/isaac/dao/IBookmarks.java
@@ -1,6 +1,6 @@
-package uk.ac.cam.cl.dtg.isaac.dos;
+package uk.ac.cam.cl.dtg.isaac.dao;
 
-import uk.ac.cam.cl.dtg.isaac.dto.users.RegisteredUserDTO;
+import uk.ac.cam.cl.dtg.isaac.dos.BookmarkDO;
 
 import java.util.List;
